add loginPage to main website
required a lot of changes to views.py and  the addition of a forms.py to hold the form information and interface it with the db.

There are also some file refactors in this commit. images have been removed from the assets folder and the assets folder has been removed