db improvements
removed the hack I made to make  entries in the student table and replaced it with a more robust receiver system. Editied the models and csv files to match up with the csv imports

also added scripts to make resetting the database a little quicker. The bash script version needs testing as I do not own a bash terminal